refactor: rename modules for clarity and update package.json description
- download/ → converter/ (better reflects ffmpeg HLS→MP4 responsibility)
- notion/ → notion-matcher/ (distinguishes from notion-uploader; this module
  only does DB discovery and class-to-page matching, not uploading)
- notion-upload/ → notion-uploader/ (symmetric with notion-matcher)

Updates all import paths in src/index.ts, src/state/index.ts, and
src/notion-uploader/index.ts accordingly.

Also updates package.json description to reflect the full three-stage pipeline.
