+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "fiap-to-notion",
   "version": "1.0.0",
-  "description": "A web scraper that authenticates, extracts PDFs and videos from my FIAP course classes, and organizes the content in my personal Notion for structured studying.",
+  "description": "CLI pipeline that scrapes FIAP course materials, converts HLS video streams to MP4, and uploads them to Notion as embedded video blocks.",
   "author": "Kauan Schaeffer",
   "license": "MIT",
   "main": "index.js",
